% Generated by roxygen2 (4.0.2): do not edit by hand
\name{alpha}
\alias{alpha}
\title{Alpha color}
\usage{
alpha(colour, alpha = NA)
}
\arguments{
\item{colour}{the colour we wanna get transparent}

\item{alpha}{the degree of transparency}
}
\value{
a new colour
}
\description{
This function gives the final color (including transparency)
}
\author{
Julien Duvanel
}

