% Generated by roxygen2 (4.0.2): do not edit by hand
\name{build_alpha_epistatic}
\alias{build_alpha_epistatic}
\title{Build alpha}
\usage{
build_alpha_epistatic(u, s, b)
}
\arguments{
\item{u}{number of non_zero coeff}

\item{s}{number of SNPS}

\item{b}{value of runif(b[1],b[2])}
}
\value{
alpha with dominant effect
}
\description{
Build alpha s.t. X = G * alpha but alpha(g_i,g_j)
}
\author{
Julien Duvanel
}

