% Generated by roxygen2 (4.0.2): do not edit by hand
\name{build_alpha_multi}
\alias{build_alpha_multi}
\title{Build alpha}
\usage{
build_alpha_multi(s, t, mode = 1)
}
\arguments{
\item{s}{number of SNPS}

\item{t}{number of traits}

\item{mode}{1 = random, 2 = disjoint, 3 = half random / half noise, 4 = half disjoint / half noise}
}
\value{
alpha
}
\description{
Build alpha multi s.t. X = G * alpha
}
\author{
Julien Duvanel
}

