% Generated by roxygen2 (4.0.2): do not edit by hand
\name{build_matrix_G}
\alias{build_matrix_G}
\title{G matrix construction}
\usage{
build_matrix_G(snps)
}
\arguments{
\item{snps}{a matrix containing n rows and p snps}
}
\value{
return a matrix of size n \times n
}
\description{
Build G matrix
}
\author{
Julien Duvanel
}

