% Generated by roxygen2 (4.0.2): do not edit by hand
\name{build_matrix_K}
\alias{build_matrix_K}
\title{Build matrix K}
\usage{
build_matrix_K(K)
}
\arguments{
\item{K}{the matrix that contain theo kinship}

\item{P}{the matrix that contain phenotypes}
}
\value{
matrix K with all relationship
}
\description{
Build the matrix K
}
\author{
Julien Duvanel
}

