% Generated by roxygen2 (4.0.2): do not edit by hand
\name{build_SNPs_matrix}
\alias{build_SNPs_matrix}
\title{Build a random Snps matrix}
\usage{
build_SNPs_matrix(n, s, snps_value = c(0, 1, 2))
}
\arguments{
\item{n}{sample size}

\item{s}{number of SNPS}

\item{snps_value}{value that SnPs can take}
}
\value{
a random SnPs matrix
}
\description{
Build a random Snps matrix
}
\author{
Julien Duvanel
}

