% Generated by roxygen2 (4.0.2): do not edit by hand
\name{build_X_multi}
\alias{build_X_multi}
\title{Build X}
\usage{
build_X_multi(M, alpha)
}
\arguments{
\item{M}{genome matrix}

\item{alpha}{matrix of traits}

\item{m}{the mode}
}
\value{
X
}
\description{
Build X using alpha and M with multi traits and mode
}
\author{
Julien Duvanel
}

