% Generated by roxygen2 (4.0.2): do not edit by hand
\name{compare_dcor}
\alias{compare_dcor}
\title{Compare DCOR and LM}
\usage{
compare_dcor(n, s, u, b, noise.sd, delta_add, delta_dom, delta_epi, snps_value,
  variable = "")
}
\arguments{
\item{n}{sample size}

\item{s}{number of SNPS}

\item{u}{number of SnPs that really explain the phenotype}

\item{b}{X = G * runif(, -b, b)}

\item{delta_add}{importance of additive effect}

\item{delta_dom}{importance of dominant effect}

\item{delta_epi}{importance of epistatic effect}

\item{variable}{the variable that vary}

\item{noise}{noise}

\item{snps_values}{the value randomly chosen for the G matrix}
}
\value{
export a pdf file
}
\description{
Compare DCOR with different kind of settings
}
\author{
Julien Duvanel
}

