% Generated by roxygen2 (4.0.2): do not edit by hand
\name{compare_dcor_multi}
\alias{compare_dcor_multi}
\title{Compare DCOR and LM}
\usage{
compare_dcor_multi(n, s, t, snps_value, variable = "")
}
\arguments{
\item{n}{sample size}

\item{s}{number of SNPS}

\item{t}{the number of traits}

\item{variable}{the variable that vary}

\item{m}{mode (see build_alpha_multi)}

\item{snps_values}{the value randomly chosen for the G matrix}
}
\value{
export a pdf file
}
\description{
Compare DCOR with different kind of settings and multi-traits
}
\author{
Julien Duvanel
}

