% Generated by roxygen2 (4.0.2): do not edit by hand
\name{compute_W}
\alias{compute_W}
\title{Build a normalized version of the genomic matrix}
\usage{
compute_W(G)
}
\arguments{
\item{G}{a genomic matrix}
}
\value{
W a normalized version of the genomic matrix
}
\description{
Build W a normalizd version of the genomic matrix
}
\author{
Julien Duvanel
}

