% Generated by roxygen2 (4.0.2): do not edit by hand
\name{DoBatchEstimation}
\alias{DoBatchEstimation}
\title{doing estimations in batch}
\usage{
DoBatchEstimation(param.list = NULL, ...)
}
\arguments{
\item{param.list}{List of parameters (containing n,p,g, etc.)}

\item{...}{arguments used by the GenerateDatasets() function.}

\item{rfunction}{from which distribution do we want to simulate data ?}
}
\value{
Nothing but export pdf in results/time_stamp[...].pdf
}
\description{
Do the simulations for each element of param.list
}
\author{
Julien Duvanel
}

