% Generated by roxygen2 (4.0.2): do not edit by hand
\name{DoEstimation}
\alias{DoEstimation}
\title{Do estimation}
\usage{
DoEstimation(get_estimate, P, V, phi, dcov = F)
}
\arguments{
\item{get_estimate}{This argument has to be a function and must return estimate's in $h}

\item{...}{arguments if needed}
}
\value{
estimate h
}
\description{
Do an estimation using the method described
in the function get_estimate (which is an argument here)
}
\author{
Julien Duvanel
}

