% Generated by roxygen2 (4.0.2): do not edit by hand
\name{estimate_heritability}
\alias{estimate_heritability}
\title{Estimate heritability}
\usage{
estimate_heritability(V, phi)

estimate_heritability(V, phi)
}
\arguments{
\item{V}{matrix built from P,K}

\item{phi}{2*Kinship}

\item{V}{matrix built from P,K}

\item{phi}{2*Kinship}
}
\value{
heritability + p.value obtained from lm

heritability + p.value obtained from lm
}
\description{
Estimate heritability

Estimate heritability
}
\author{
Julien Duvanel

Julien Duvanel
}

