% Generated by roxygen2 (4.0.2): do not edit by hand
\name{estimate_heritability_dcor_H2}
\alias{estimate_heritability_dcor_H2}
\title{Estimate heritability with dcov / LN version}
\usage{
estimate_heritability_dcor_H2(V, phi)
}
\arguments{
\item{V}{matrix built from P,K}

\item{phi}{2*Kinship}
}
\value{
heritability + p.value obtained from lm
}
\description{
Estimate heritability with dcor
}
\author{
Julien Duvanel
}

