% Generated by roxygen2 (4.0.2): do not edit by hand
\name{estimate_heritability_dcov}
\alias{estimate_heritability_dcov}
\title{Estimate heritability}
\usage{
estimate_heritability_dcov(V, phi)
}
\arguments{
\item{V}{matrix built from P,K}

\item{phi}{2*Kinship}
}
\value{
heritability + p.value obtained from lm
}
\description{
Estimate heritability
}
\author{
Julien Duvanel
}

