% Generated by roxygen2 (4.0.2): do not edit by hand
\name{filter_data}
\alias{filter_data}
\title{Filter P}
\usage{
filter_data(P, phi.matrix)
}
\arguments{
\item{P}{the matrix that contain phenotypes}

\item{phi.matrix}{the matrix containing the values for phi_pq}
}
\value{
P and phi filtered and without NA
}
\description{
Filter P to keep the correct phenotypes (and remove NA)
}
\author{
Julien Duvanel
}

