% Generated by roxygen2 (4.0.2): do not edit by hand
\name{filter_P_by_id}
\alias{filter_P_by_id}
\title{Filter P}
\usage{
filter_P_by_id(P, id)
}
\arguments{
\item{P}{the matrix that contain phenotypes}

\item{phenotype}{ID}
}
\value{
P filtered and without NA
}
\description{
Filter P to keep the correct phenotypes (and remove NA)
}
\author{
Julien Duvanel
}

