% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_available_phenotypes}
\alias{get_available_phenotypes}
\title{List of available phenotypes}
\usage{
get_available_phenotypes(P)
}
\arguments{
\item{P}{the matrix that contain phenotypes}
}
\value{
row id and phenotype name
}
\description{
Obtain a list of availale phenotypes
}
\author{
Julien Duvanel
}

