% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_frequency}
\alias{get_frequency}
\title{Get frequency of minor allele}
\usage{
get_frequency(G, snp)
}
\arguments{
\item{G}{genomic matrix}

\item{snp}{a snp (index)}
}
\value{
frequency of the minor allele
}
\description{
Get frequency of the minor allele for the SNP snp
}
\author{
Julien Duvanel
}

