% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_P_row_from_full_id}
\alias{get_P_row_from_full_id}
\title{Get the row ID of P according to the full_id}
\usage{
get_P_row_from_full_id(full_id, P)
}
\arguments{
\item{full_id}{the id that we want to match}

\item{P}{the matrix we want to find the row (must contains P$full_id)}
}
\value{
row number of P$full_id == full_id
}
\description{
Generate a correlation matrix with parameters in argument.
Notice that corr can be a vector.
}
\author{
Julien Duvanel
}

