% Generated by roxygen2 (4.0.2): do not edit by hand
\name{GetLegendFromGgplot2}
\alias{GetLegendFromGgplot2}
\title{Get ggplot's legend}
\usage{
GetLegendFromGgplot2(a.gplot)
}
\arguments{
\item{a.gplot}{a ggplot object}
}
\value{
legend of a ggplot's object
}
\description{
Get legend from a ggplot's object
}
\author{
Julien Duvanel
}

