% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PlotHeatmapCorrelationMatrix}
\alias{PlotHeatmapCorrelationMatrix}
\title{Plot the heatmap of X's correlation matrix}
\usage{
PlotHeatmapCorrelationMatrix(X, correlation = FALSE)
}
\arguments{
\item{X}{original matrix}

\item{sparsity}{choosen groups}
}
\value{
Plot the heatmap
}
\description{
Plot the heatmap of X's correlation matrix
}
\author{
Julien Duvanel
}

