% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PlotScatterMethodsBis}
\alias{PlotScatterMethodsBis}
\title{Scatter plot methods}
\usage{
PlotScatterMethodsBis(full.res, heridiff.file, editplotname.file,
  pdf.path = "")
}
\arguments{
\item{full.res}{results obtained after estimating heritability}

\item{heridiff.file}{file containing estimates given by Chiara}

\item{editplotname.file}{file containing PlotGroup (needed to plot group)}

\item{pdf.path}{where the pdf are exported}
}
\value{
nothing but export a pdf
}
\description{
Compare methods via ScatterPlot
}
\author{
Julien Duvanel
}

