% Generated by roxygen2 (4.0.2): do not edit by hand
\name{product_snps_alpha}
\alias{product_snps_alpha}
\title{Product between mat and alpha for additive effect}
\usage{
product_snps_alpha(M, alpha)
}
\arguments{
\item{M}{a matrix with s columns}

\item{alpha}{a vector}
}
\value{
product of mat * alpha
}
\description{
mat * alpha but alpha(mat)
}
\author{
Julien Duvanel
}

