% Generated by roxygen2 (4.0.2): do not edit by hand
\name{product_snps_alpha_dominant}
\alias{product_snps_alpha_dominant}
\title{Product between mat and alpha for dominant effect}
\usage{
product_snps_alpha_dominant(M, alpha)
}
\arguments{
\item{M}{a matrix with s columns}

\item{alpha}{a matrix with 3 rows and s columsn}
}
\value{
product of mat * alpha
}
\description{
mat * alpha but alpha(mat)
}
\author{
Julien Duvanel
}

