% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ReadGRMBin}
\alias{ReadGRMBin}
\title{R script to read the GRM binary file}
\usage{
ReadGRMBin(prefix, AllN = F, size = 4)
}
\arguments{
\item{prefix}{name of the files}

\item{size}{.}

\item{allN}{.}
}
\description{
R script to read the GRM binary file
}
\author{
plink website
}

