% Generated by roxygen2 (4.0.2): do not edit by hand
\name{TraitGenerator}
\alias{TraitGenerator}
\title{Generate fake traits}
\usage{
TraitGenerator(plink.file, export.path, N = 20,
  sbatch.file = "ExtractSNPs.sbatch")
}
\arguments{
\item{plink.file}{where are stored plink files}

\item{export.path}{where to export RData fake traits}

\item{N}{number of SNPs to extract from genome}

\item{sbatch.file}{where the sbatch file will be saved}
}
\value{
nothing
}
\description{
Extract randomly a few SNPs from genome and then generate fake traits
}
\author{
Julien Duvanel
}

