% Generated by roxygen2 (4.0.2): do not edit by hand
\name{vplayout}
\alias{vplayout}
\title{Wrapper for viewport}
\usage{
vplayout(x, y)
}
\arguments{
\item{x}{number of rows}

\item{y}{number of columns}
}
\description{
Wrapper to easily use one of the viewport option
}
\author{
Julien Duvanel
}

