% Generated by roxygen2 (4.0.2): do not edit by hand
\name{WriteSBATCH}
\alias{WriteSBATCH}
\title{Write SBATCH}
\usage{
WriteSBATCH(plink.file, sbatch.file, character.snps, export.path,
  datetime.stamp)
}
\arguments{
\item{plink.file}{where are stored plink files}

\item{sbatch.file}{where the sbatch file will be saved}

\item{character.snps}{list of SNPS (character separated with coma)}

\item{export.path}{where to export RData fake traits}

\item{datetime.stamp}{time stamp}
}
\value{
nothing
}
\description{
Write SBATCH
}
\author{
Julien Duvanel
}

